+++ src/Request/APIKey/CreateAPIKeyRequest.php
@@ -22,12 +22,12 @@ class CreateAPIKeyRequest
     public function toArray(): array
     {
         $data = [
-            'project_id' => $this->projectId,
+            'projectId' => $this->projectId,
             'name' => $this->name,
         ];
 
         if ($this->userId !== null) {
-            $data['user_id'] = $this->userId;
+            $data['userId'] = $this->userId;
         }
         if ($this->scopes !== null) {
             $data['scopes'] = $this->scopes;
@@ -36,10 +36,9 @@ class CreateAPIKeyRequest
             $data['ttl'] = $this->ttl;
         }
         if ($this->allowedIps !== null) {
-            $data['allowed_ips'] = $this->allowedIps;
+            $data['allowedIps'] = $this->allowedIps;
         }
 
         return $data;
     }
 }
-